// types/sources.ts
export type SourceId =
  // Reddit
  | 'reddit_til'
  | 'reddit_science'
  | 'reddit_damnthatsinteresting'
  | 'reddit_askhistorians'
  | 'reddit_eli5'
  | 'reddit_machinelearning'
  | 'reddit_physics'
  | 'reddit_biology'
  | 'reddit_space'
  | 'reddit_natureismetal'
  | 'reddit_humansbros'
  | 'reddit_madesmile'
  | 'reddit_upliftingnews'
  | 'reddit_standup'
  | 'reddit_writingprompts'
  | 'reddit_books'
  | 'reddit_movies'
  | 'reddit_economics'
  | 'reddit_wsb'
  | 'reddit_philosophy'
  | 'reddit_etymology'
  | 'reddit_askscience'
  | 'reddit_dataisbeautiful'
  | 'reddit_architecture'
  | 'reddit_earthporn'
  | 'reddit_writing'
  // Wikipedia
  | 'wikipedia_top'
  | 'wikipedia_onthisday'
  | 'wikipedia_unusual'
  // APIs & feeds
  | 'arxiv'
  | 'hackernews'
  | 'nasa_apod'
  | 'inaturalist'
  | 'aeon'
  | 'quanta'
  | 'longreads'
  | 'lithub'
  | 'marginalian'
  | 'naturepicks'
  | 'phys_org';